tests/requirements.txt
docker>=2.4.2
pytest>=3.2.1
pytest-splinter>=1.8.5
requests>=2.18.2
